Rubberduck issues
It seems that the Rubberduck asseerts it broken at the moment.  Hence lots of tests were flagged as passing when In fact the results of the test indicated that  they should have failed.  VBALin has been exported to twinBasic and code for failing tests corrected before importing back into VBA.

In addition, I was forgetting to update the vbattributes for the source code after Rubberduck project export (for PredeclaredId the VBCreatable attribute needs to be set to True).  This has now been done for the exported source files.