Maths.Permutations, ArrayOpSplat method, Fmt array error
Converted the Maths permutations method to be a function with an optional size parameter rather than returning the permutations by an out parameter.

Moved the fmt splat operator to be a more general purpose ArrayOp method.

Resolved a but whereby passing fmt.Dbg/Text an array would result in just the first parameter of the array being printed.