Revised AssertExactSequenceEquals to use 'or' rather than 'and' un the loop statement.
Feedback highlighted an potential shortcoming of AssertExactSequenceEquals in the use of 'And' in the Loop statement which could lead to a false positive.  Changing to And ensures that we don't get a flase positive.

Also some minor updates for issues picked up by twinBAsic that are not flagged by VBA or Rubberduck (e.g. End Function vs End Property' and similar.