Added Permutations class and wqetos crypto module
The permutations module calculates Permutations using Heap's algorithm.  This version calculates the permutations using a set of longs.  This set of longs is then used as a set of indexes with which we can lookup the true values.  This allows the perutations methods to be fast and to work for any set of variables.

Wqeto's Crypto module was added as it is faster the net crypto services, however, for some reason It runs into a overflow issue.