Added Join method to Seq and Kvp classes.  Added Inner method to Strs class
Added a Join method to Seq and Kvo classes which returns a string representation of the Items stripped.  The string is generated by Fmt.Text so the first and last characters are stripped to remove the enclosing { and } object delimiters.

Introduced Inner method for Strs class.  Inner removes the specified number of characters from the left and righthand sides of a string.