Added ISeq interface, Seq.Transpose method, fold function fnMaxNum and udated rdCountIt reducer classs.
Interested in consolidating 'sugar' methods of Seq and Kvp classes which tend to have the same generic code.   To do this I added an ISeq interface using Rubberduck and added the interface calls to each seq class.  Initial use of the interface in the new Transpose method showed that it was not working as expected.

Added Transpose method which transposes a seq of seq.

added Folf function fdMaxNum

Updated the rdCountit class to make the comparer provided at initialisation an option, then updated the code so that the number of items in the seq is returned if the Comparer is 'nothing'.