Updated ISeq bits missed in last update.  Chanfes seq Add byref to Byval
Missed updating the Iseq methods in the Seq classes on last update.

The seq Add method passed the parameter by ref.  This was found to create an issue when adding the parameter to a seq as the byref meant that the string previously added to the seq reflected the current value and not the added value.  This behaviour wqs corrected by chaning to ByVal.  Proved in AoC 2022 Day 06