Updating of Reducers to take a paramarray as initialisation
IN developing function to use with IFold is became apparent that it would be useful to be able to return the index rather than the accumulator.  Consequently the rdFoldl and rdFolr methods were updated to return an array comprised of accumulator and the index at which the accumulator last changed.

It also became apparent that certain aspects of reducers/comparers  would be clearer and more flexible if the initalisation input was a ParamArray.  Interfaces work by matching the method types and not parameter names, consequently a descriptive name can be used in the specific folder i.e. see cmpIn  cmpNotIn