Introduce Allow module. Updated cmp functors to use paramarrays. Initial thoughts on a NUmber class
The current Guard methods in the Guard Module are confusing to use because of the double negatives.  Consequently I've introduced the Allow class which take the opposite approach to a negative test.  I.e. it only continues if the condition is true.

The cmpIn and cmpNoy in classes require to limits to be applied, so rather than getting upto code jiggery pokery for these two classes to get them to comply with IComparer interface, the functors have been refactored to use ParamArrays for the initalisation inputs in the Deb method.

Sometimes we want to allow a number but not a variant.  First thoughts on a number class are included.  It would make better sense to make the number class more generic by making the number class deal with complex numbers.  That's a ToDo.

Other minor updates due to bug fixing