Updated Stringifier to allow type information to be included in the output string.
Added the ability for Stringifier to include Type information in the output string.  The Type information required can be specified at initialisation of via the WithTypes method.

The options available are
None
Inner: Type prefixes for Array and Container Types are not included
Outer: Type Prefixes inside container types are not provided
All:  Complete Type information is provided.